user/requirements.txt
django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt
drf-spectacular==0.25.1
pygments==2.14.0
gunicorn==20.1.0
json-log-formatter==1.1
django-cors-headers
django-otp
django-two-factor-auth
qrcode
httpie